feat: create label option in create issue modal (#281)
