+++ apps/app/components/labels/index.ts
@@ -1,3 +1,4 @@
+export * from "./create-label-modal";
 export * from "./create-update-label-inline";
 export * from "./labels-list-modal";
 export * from "./single-label-group";
