+++ apps/app/components/states/index.ts
@@ -1,4 +1,4 @@
 export * from "./create-update-state-inline";
-export * from "./create-update-state-modal";
+export * from "./create-state-modal";
 export * from "./delete-state-modal";
 export * from "./single-state";
